clean up repo structure and README
