--- app_2.py
@@ -1,10 +1,0 @@
-from flask import Flask 
-
-app = Flask(__name__)
-
-@app.route('/')
-def hello_world():
-	return 'Hello, World!'
-
-if __name__ == '__main__':
-	app.run()
