--- flask_3.py
@@ -1,10 +1,0 @@
-from flask import Flask, render_template 
-
-app = Flask(__name__)
-
-@app.route('/hello')
-def hello():
-	return render_template('hello.html')
-
-if __name__ == '__main__':
-	app.run()
