+++ templates/hello.html
@@ -1,10 +1,12 @@
 <!DOCTYPE html>
-<html>
+<html lang="en"> <!-- tells browser this page is in English -->
 <head>
+	<meta charset="UTF-8"> <!-- handles special characters -->
+	<meta name="viewport" content="width=device-width, initial-scale=1.0"> <!-- mobile friendly -->
 	<title>Hello, Flask!</title>
 </head>
 <body>
-	<h1>Hello, Flask!</h1>
+	<h1>Hello!</h1>
 	<p>Welcome to my Flask app!</p>
 </body>
 </html>
